refactor: Adjust import paths for authentication actions
